pyproject.toml
[tool.poetry]
name = "kursovaja4"
version = "0.1.0"
description = ""
authors = ["Mihkel Metsa <mihkel992@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
